Partially refactor into data_structures.c
