+++ R/nvimcom/src/apps/nvimrserver.c
@@ -1371,43 +1371,14 @@ void update_pkg_list(char *libnms) {
     }
 }
 
-ListStatus *search(const char *s) {
-    ListStatus *node = listTree;
-    int cmp = strcmp(node->key, s);
-    while (node && cmp != 0) {
-        if (cmp > 0)
-            node = node->right;
-        else
-            node = node->left;
-        if (node)
-            cmp = strcmp(node->key, s);
-    }
-    if (cmp == 0)
-        return node;
-    else
-        return NULL;
-}
-
-ListStatus *new_ListStatus(const char *s, int stt) {
-    ListStatus *p;
-    p = calloc(1, sizeof(ListStatus));
-    p->key = malloc((strlen(s) + 1) * sizeof(char));
-    strcpy(p->key, s);
-    p->status = stt;
-    return p;
-}
-
-ListStatus *insert(ListStatus *root, const char *s, int stt) {
-    if (!root)
-        return new_ListStatus(s, stt);
-    int cmp = strcmp(root->key, s);
-    if (cmp > 0)
-        root->right = insert(root->right, s, stt);
-    else
-        root->left = insert(root->left, s, stt);
-    return root;
-}
-
+/**
+ * TODO: Candidate for data_structures.c
+ *
+ * Description:
+ * @param s:
+ * @param stt:
+ * @return
+ */
 int get_list_status(const char *s, int stt) {
     if (listTree) {
         ListStatus *p = search(s);
@@ -1420,6 +1391,12 @@ int get_list_status(const char *s, int stt) {
     return stt;
 }
 
+/**
+ * TODO: Candidate for data_structures.c
+ *
+ * Description:
+ * @param s:
+ */
 void toggle_list_status(const char *s) {
     ListStatus *p = search(s);
     if (p)
